Extract file transfer routes
